Move translations to src folder and create/writeLangIds to .bin
