--- .bin/create-lang-ids.js
@@ -1,4 +1,0 @@
-/* eslint-disable */
-const { writeLangIdSubstringMap } = require("../src/createLangIds");
-
-writeLangIdSubstringMap();
