+++ .bin/writeLangIds.ts
@@ -1,0 +1,4 @@
+/* eslint-disable */
+const { writeLangIdSubstringMap } = require("./createLangIds");
+
+writeLangIdSubstringMap();
