+++ .gitignore
@@ -1,6 +1,6 @@
 node_modules/
 dist/
-translations/*.ts
+src/translations/*.ts
 *.tgz
 yarn-error.log
 .eslintcache
