+++ .prettierignore
@@ -1,4 +1,4 @@
 node_modules
 dist
 coverage
-translations/*.yml
+src/translations
