+++ tsconfig.eslint.json
@@ -1,4 +1,4 @@
 {
-  "include": ["./**/*.test.ts"],
+  "include": ["./**/*.test.ts", "./.bin/**/*.ts"],
   "exclude": ["*.js", "*.ts"]
 }
