+++ tsconfig.json
@@ -26,5 +26,5 @@
     "inlineSources": true
   },
   "exclude": ["node_modules", "src/**/*.test.ts"],
-  "include": ["src/**/*.ts", "translations/**/*.ts"]
+  "include": ["src/**/*.ts"]
 }
